build: Add target for simulator binary.
This requires all of the automake rules as well as pulling in the
pthread headers and libraries.

The structure of the build mimicks the visual studio build as much as
possible. The static libraries libplatform & libtpm have a cyclic
dependency so the LDADD line for the simulator binary is a bit weird. It
may be better to just put all of the objects into one archive.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>