tpm: Copy TPMA_MERMANENT to UINT32 destination with memcpy.
This is a hack to work around assumptions about Visual C++ specific
behavior in the compiler. Even though the required number of bits are
defined we can't subvert the type system and cast it to an arbitrary
type before dereferencing it. This breaks strict aliasing rules enforced
when we use -Wall & -Werror with gcc.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>

fixup into 90c056071d109c0235c5f964c5a06a6303bef15b