tpm: Use OpenSSL environment to determine proper RADIX_BITS.
Expecting users to change code in order to build for 64bit environments
isn't very user friendly. In the case of the RADIX_BITS we can make this
easier on users (and automated builds) by using the openssl bn.h header
to do the detection for us. This is very similar to the sanity checks in
the `TpmToOsslMath.h`.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>