tpm: Use OpenSSL environment to determine proper RADIX_BITS.
Expecting users to change code in order to build for 64bit environments
isn't very user friendly. In the case of the RADIX_BITS we can make this
easier on users (and automated builds) by using the openssl bn.h header
to do the detection for us. This is very similar to the sanity checks in
the `TpmToOsslMath.h`.

For the windows build we must add the openssl include path for the
Platform module since it includes the Implementation header. We also
must ensure __STDC_WANT_SECURE_LIB__ is defined before including bn.h
to prevent a warning from the MS compiler from producing a warning. We
define this macro to 0 to prevent the inclusion of non-standard
functions.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>